Add a list of TODO tasks to the project.
The file TODO contains the list of tasks
to be done. Whenever a task is finished
by one of the group members, this person
should write -(REVIEW) in front of it.
This way other group member can review
the work done and remove the task from
the list. Anyone can add more tasks to
the list whenever a new demand is identified.