Schedule
As requested by Leandro I placed the schedule in the repo. Feel free to
make changes. Also, I think it would be a good idea to leave comments
on this document. The comments could include status of a certain task,
where you are on a particular task, if you came across any errors while
test, etc. Let me know if the file type doesn’t work for you.