Add Comments
There was a mistake in the addition of the schedule to the repo. 1 copy
of the schedule was placed in the docs folder while another was placed
in the main directory. The schedule in the docs folder is the correct
file, and this should be the one used because it is the most
up-to-date. The previous version in the main directory will be deleted.