Revert "Fixed exponential notation floats"
This reverts commit a7e7fa9fa709eae2d0472d1fb25c0752f8fd64af.
Exponential notation is not allowed anymore.
The reason is it is hard to read floats
using exponential notation from the keyboard
since SAM doesn't allow it. In this case we
would have to read floats as strings and
parse the strings in SAM assembly. If we decide
to allow this notation only for the programmers,
which I think doesn't make sense, we can do that
by reverting this commit. In that case, please make
sure to fix the scanner. According to the MnC manual,
on page 4, the signal(+/-) after the E/e is mandatory.